MessageScheduler Fixed
Fixed an issue with the timing of events, specifically, the docker image
runs on UTC time so I had to make the time string include the UTC
offset. (e.g. -0400 for EDT)
API is basically complete, just need to add a linting function for the
time param, and return an error as necessary.