Added connection between Scheduler and API
Added a mutually exclusive parameter to the Scheduler endpoint that
specifies an endpoint to query that can be used instead of the 'msg'
parameter.
The current implementation only allows users to send 'GPS' as an
endpoint to check and generate a message from, just for the sake of
timing on the project; additional endpoints should not be difficult to
add from this base.