Remove Invalidated Twilio credentials
I received notification from Twilio that the credentials were publicly posted and eventually tracked it down to this file.  

A new token was generated, so the old one (which will still be part of the git history) no longer works.  But I wanted to
change this file so that if someone finds it later they don't try to use it elsewhere...